beginning work on target screen
this is going to be complicated.
instead of having a target screen create "targets" from visible enemies, i think i will have every formula be manually aimed. this should give a more dynamic feel to the combat, requiring additional player judgment.
the "target screen" is its own canvaslayer, which runs while the rest of the game is paused. the targeting reticle will change depending on the formula used. "targets" will be created at enemy positions and added to the target screen. the first issue i've encountered is getting the enemy's position and applying it to the target(s). i believe the issue is because of different canvasLayers. the target position is always relative to the screen position. i will need to think of another way to do this.