targeting screen implemented and linked to spellcasting
basically, instance a FormulaTarget scene in the formula scene you wish to have targeting, and play with the code a bit. remove the _ready() function from the formula and replace it with a function you can run manually, start(), allowing us to cancel the casting of the formula from the target screen.
the target screen creates a node called "Targets" as a child of World. it then iterates through all enemies on-screen, creates "EnemyTargets" in their positions, and adds them as children to World/Targets. these are the circles that indicate enemies on the targeting screen.  they are of type StaticBody2D and exist on the "Target" collision layer. the player's formula reticle looks for bodies on the "Target" collision layer and turns red if a target is inside the reticle. the target turns red as well. this feature can potentially be expanded, showing different effects based on the formula used. ice formula? maybe a crystalline targeting reticle, etc. target weak to the current formula? maybe change the way its target displays. resistant? ditto, etc.
the target screen can be cancelled out by either hitting the "examine" or "roll" keys. if cancelled using the roll key, the Player will perform a roll in the direction held or a backstep if no direction is held. at first, i wanted to get rid of this happening, but i realized it could be beneficial for the player if they had second thoughts about casting their formula and instead decided to quickly relocate. so i decided to keep it, and have the "examine" button cancel out with no action performed. this was done by adding an additional check to examining: a "casting" boolean was added to the player. it is set to true upon using a formula, and set back to false 0.35s later, preventing the "examine" button from inadvertently opening up a dialog box.
if a formula is cancelled out from the target screen, its node is deleted. thus, before we add spell XP and take away ingredients, we check !is_instance_valid(formula) on the formula cast, and return out of the function if it's not. otherwise, spell XP is added and ingredients are used up.
i'm not sure how scale-able the current setup is for spellcasting. i will need to come up with various types of spells with a variety of targeting methods (different size/shape of AoE, different cast time, different effects, etc.) to test thoroughly, but so far it seems to be working OK.