organization / bugfix
putting all formulas into specific subfolders in the assets/Player/Abilities folder.
fixed a bug where if a Player zoned with a formula equipped for which they did not have any the necessary ingredient(s), the ingredient display in the formulaUI would display Null.
next i'm thinking of writing a function for healing to be able to display the exact heal values on screen similar to damage values. the simplest way to do this would be to incorporate it into the damage dealing functions themselves. the actual displaying of damage values takes place in the hurtbox scripts for the player & enemies (which are discrete from one another). before that, damage calculation occurs in the Global.damage_calculation method, for both player & enemies.
now, to get heal values to display, i will have to think about and re-do how health values are applied. as of now, the heal spell doesn't go through any global damage calculation method. heal potency is calculated on a cast-by-cast basis on the Heal scene itself:

heal_amount = (10*FormulaStats.Heal[1]) * PlayerStats.magic_mod*PlayerStats.magic_mod

10 is the "base_potency", which is multiplied by the level of the Heal formula. this is then multiplied by the square of PlayerStats.magic_mod, which is 1 + magic/32. this determines the heal_amount, which is then applied directly to PlayerStats.health.
so a problem i can see without even doing anything is that the Heal formula increases health over time: it is not a lump value being applied to PlayerStats.health. rather, it is 1 health being applied every frame until a threshold is reached, heal_amount. the lazy solution would be to just display heal_amount, either at the beginning or at the end of the Heal process. some thoughts...
right now, i don't believe the heal process can get interrupted, even if you get hit. if you start a heal spell, you will get healed for its full value, even if you get hit. i might have healing be interruptable to incentivize the player to wait/find the best healing opportunities as right now it's possible to simple "tank" through any damage received while healing. this can be intentional or not - we will have to see how it feels.
i think as of now, coming up with a way to just display the total_healed value will be a good start, and then other features can be worked on/modified from there.