bugfixing target screen
a major bug i noticed was the sometimes, likes less than 10% of the time, cancelling out of the target screen would still take away ingredients and award formula XP. this was because i was using the function is_instance_valid(formula) to check after when the player's CastTimer had elapsed if the formula was deleted or not. the issue was that sometimes the "formula" object reference had been replaced by another random node. this is a problem with the Godot engine itself so i needed to find a workaround.
basically, now, after the CastTimer times out, i check if the World node has a child of the same name as the formula cast, and only remove ingredients/add XP if so.
i also added a feature to snap to enemies in range while in the target screen. press the "next" button (R) to cycle through enemies in range.