minor fixes
changed text display of the pickup popups (which also show spell level ups): uses a bigger font and repositioned a little higher to be more centred.
also fixed a big where scrolling through enemies with some out of range would skip certain enemies on the next pass through. this was because the "count" was not being updated properly when the function finds one out of range: it would remove the enemy from the enemies array, but the count would still increase, meaning whichever enemy got placed in the old index would get skipped over. this would happen for each enemy removed from the array, resulting in an extremely unpredictable enemy order through which we scroll.