created test boss
it's likely that each boss is going to require a unique script. for now, decided to instantiate a boss who still uses the Enemy script, with some tweaks.
what makes the boss unique:
for now, not much, other than there being no background music until the boss is on screen, at which point the boss health bar is shown. it works the same as enemy health bars, pretty much, other than the fact that it is on its own CanvasLayer, which is itself a child of the boss. of course, it will need a unique function for when it dies - something like, play boss dead scene, with maybe big flash/explosion that slowly dies down, then victory pose/fanfare - after which reward XP/loot/whatever.
i expect bosses to have some pretty complex behaviour, down the line. it makes sense to create boss "arenas", or rooms/scenes that are specifically for fighting a boss, that, once entered, you can't leave. this is actually how secret of mana did it, from what i remember: when a boss starts, you are moved to a new area, and moved from that area when a boss is defeated - EVEN IF the area you come from/go to looks the exact same. it's just a "boss arena version" of that area with all of its other scripts disabled, presumably. obviously this is going to take a lot experimentation to see what works.