re-wrote healing functions
the code in place will definitely get messier when more elements are added to damage/formula potency calculation - things like weaknesses and such - so i wouldn't consider the current version very scaleable or final, but it works for now.
i wanted to keep the value healed to go up in a gradual way (as in +1hp/frame after casting until heal_amount is reached as opposed to a lump sum healed at once), but i also wanted to use a hitbox to detect bodies in the heal radius for a few reasons, but primarily to keep things consistent with other formulas/things that do "damage", where "damage" can be either positive (actual damage) or negative (healing).
when a hitbox enters the player's hurtbox, checks if it is a formula. since the only hitbox programmed into the game that is a formula that targets the player's hitbox is the Heal, the only thing that happens after a hitbox detected to be a formula enters the PlayerHurtbox is to display the heal amount. normally, damage calculation would take place here, which takes into account various stats on both the owner of the hitbox and the owner of the hurtbox. however, Heal potency is calculated inside the Heal scene itself, and thus the only thing that happens when a formula hitbox enters the player hurtbox is to display the amount healed. displaying a heal popup as opposed to a damage popup runs a different function, which is not ideal. the reason this is is because damage values are converted to strings before passing them as arguments to the instantiation of damagePopup displays. this is because the damagePopup will always interpret damageDisplay as a string, because sometimes the argument recieved is "Miss!". i wonder if it would make more sense to have Miss be its own function, and to pass damage values to the popups as integers rather than strings, so that the damagePopup itself can interpret whether damage was positive or negative, so as to either subtract or add health from/to the player, respectively.