working on making pause screen menu items into interactable buttons
doesn't actually do anything but play sounds for now. just trying to program the menus. the status section was easy enough to hard-code in the buttons, but the other categories are more confusing because the buttons will have to be instantiated depending on the formulas/ingredients found on the player, and then have their focus neighbours connected based on the positions they were created.