some clean-up and fixes
got rid of every map scene's dependency on having a FadeIn.tscn scene in the editor sceneTree on running. now, the World script - which we shall assume will be attached to every map screen in the game - instantiates its own FadeIn via code on _ready().
additionally, got rid of the need for the "Overlay" CanvasLayer. originally, this is where the LevelUpScreen displayed, because the TweenGreyScale would dim and fade out the colours of every layer below it, leaving the levelUpScreen the only thing still lit & coloured. now, the TweenGreyScale is instantiated and added BEFORE the level up screen is instantiated to the GUI CanvasLayer (1), after which the levelUpScreen is added to the same layer, GUI (1). this gets rid of needless complexity and could potentially prevent some headaches in the future.
various other minor adjustments as well.