adjusted sword values
instead of having sword damage based entirely on STR (violent nature), it now has a base damage (4) which is incremented by each point put into STR. this allows for the base damage to be adjusted in the future (different weapons? weapon upgrades?) also increased the knockback a bit and decreased the randomness a bit.