beginning re-work of status effects
previously, the status effects that are in the game are not only hard-coded to the player, but also badly implemented. now that i have more knowledge and experience creating re-usable & modular systems, i want to create a rigid and expandable/scaleable status effect system that works for both Player(s) and enemies.