poison scene working on player & enemies
created a new singleton called StatusHandler, which takes 2 arguments: status (String), and body.
hitboxes can have a "status" variable, which is a String. when an object (player or enemy)'s hurtbox is entered by a hitbox, checks if that hitbox has the "status" variable, and if it does, passes it as an argument to StatusHandler, along with itself (body). StatusHandler then matches the received argument with which status to instantiate. it is then added as a child to the body. if the body already has a child node with the same name as the status being applied, returns out of the function (nothing happens).
currently, the Poison parameters are hard coded, but this can easily be changed. the parameters are potency and duration, which currently are 1 and 45. each time the PoisonNotice animatedSprite animation finishes, runs the poison_tick() function on the Poison scene, which decrements the health of the body to which poison is applied by potency (1) and displays a damage popup in purple, indicating poison damage. poison_tick() also increments the 'count' by 1; if count = duration when the poison animation finishes, it gets deleted. it will also get deleted if when it was told to poison_tick(), the body's health was less than or equal to 0.
so basically, duration is the number of ticks, and potency is the damage of each tick. the duration of each tick is currently tied to the animation. i will decide later if this is appropriate or not. it might be better to create a longer animation to increase the period between poison_ticks, or to have the function called by a Timer's timeout signal - which would allow tick duration to be modified. but for now i like that damage ticks are tied to the poison bubble "pop".
also there is not yet any calculation for applying status. i figure including a chance to resist would be appropriate. for the player, perhaps resist chance can be tied to the Resilience stat.