working on status application
if a hitbox is capable of bestowing a status effect, the hitbox will have the variable "status" which is an array with 2 indexes. index 0 is a string of the name of the status, eg. "Poison", and index 1 in a float of the chance of bestowing that status, eg. 0.5 for 50%.
i think the Resilience stat should contribute to preventing unwanted status effects. i will come up with a formula to check for that.