status resistance implemented / other adjustments
for now, status_resistance = defense/128, which is less than a single percentage point per point in Resilience. at 64 Resilience, the Player has a natural 50% resistance to status effects.
when an attack with a status effect hits a target, a randf is checked against the status_chance of the attack minus the status_resistance of the target.
for example...
if a bat's attack carries the Poison status with a chance of 50%, its hitbox will have this variable:
     bat.hitbox.status["Poison", 0.5]
now the Bat attacks our Player, whose Resilience stat is at 10. status_resistance = defense/128, so in this case, it is 0.078125. this number is subtracted from bat.hitbox.status[1], resulting in our final status application chance, 0.421875, or ~42%.
a randf is checked against this number, and if it is less than the final status application chance, the status is applied. otherwise, it is resisted.
a Resilience stat of 64 provides a natural status_resistance of 0.5, or 50%. therefore, if an enemy attack has 50% base chance of inflicting poison, and the player has 64 or more Resilience, then there is a 0% chance that attack will inflict poison. if an attack has a 100% base chance of inflicting poison, vs. a Player with 64 Resilience, there is a 50% chance poison gets applied.
i haven't tested it with higher numbers, but i assume that giving status[1]  a value of anything higher than 2 will result in a nigh-unresistable status effect, since the base chance would be 200%. a Player would need to have 128 or more points in Resilience to even have a fraction of a chance to resist a status with a base chance of 2.0.
right now, for testing purposes, the sword attack has a poison chance of 0.25 or 25%, and bats have a poison chance of 50%. the wolf has a status_resistance chance of 0.2; other enemies, 0.
also added popup messages for status effects and crits, among other adjustments.