working on fixing stun status
stuns seem to work properly for enemies, for the mostpart.
stun time currently is hard-coded. if an enemy gets stunned, it is stunned for 2 seconds and cannot break out of the stun before then, even if hit. if another stun tries to get applied, it is instead ignored; a single status cannot stack nor refresh. eg. if a stunned enemy is hit with a stun, nothing happens, and the enemy will break out when the first stun's timer times out.
for the player, things are a little more complicated, because when HIT and staggered, the player goes into the HIT state, removing them from the stun state. this is a complicated problem for a number of reasons.
i want the player to be able to be knocked out of stun on getting hit while stunned. if this happens, another stun should not be applied. the problem lies in the order of execution, somehow.
the first thing that happens when a hit is confirmed on the Player is to check the area hitting the player for a status effect by running it through the StatusHandler. if the StatusHandler detects an instance of the status already on the body, returns out of the function, doing nothing.
then, damage/crit/stagger calculation occurs as normal.
finally, hit_damage() is run, applying damage, and a check as to whether to chance the player state to HIT.
the issue is: at the end of the HIT state, i don't know how to check whether a stun as been applied ON THE HIT that sent the player into that HIT state, or whether the hit should INTERRUPT a stun that was ALREADY THERE BEFORE THE PLAYER GOT HIT.
basically, the HIT state needs to know to which state the Player should go to (MOVE or STUN) when the hit animation finishes. if the player was already stunned before the hit, the stun should be interrupted, and the player moved to the MOVE. if the hit CAUSED the stun, the player should be put into the STUN state.