adjustments
the poison icon does not disappear on death. this is because it only deletes itself when its value == max_value. its value is set when the Poison status node itself emits a signal. the poison status gets deleted when health gets to zero, and no signal is emitted. there are ways to fix this, especially since the way death works currently is NOT how death will function in the complete game, so i'm not worried about it. i figure on death, parse through all statuses & icons and just delete them outright.