modifying status system / added Regen status
status deletion is now based on time duration rather than a count number. this way, the timer of the actual status and the timer of the status icon can operate independently. when a status is removed, it emits a signal that it's been removed, which is listened for by the status icon, which it then uses to delete itself, if heard.
haven't tested because i have to run but in theory should work. next, i would like Regen & Poison to cancel each other out: if Poison is cast on a Regen'd body, then the Regen gets removed, and vice versa. also, unlike debuffs, i would like buff durations to be refreshed when they are hit with a buff they already have on them.