poison/regen cancel each other out
if regen is cast on a body that has a Poison node as a child, that child is deleted, and vice versa.
also not sure if i want status changes to be displayed as messagePopups, damagePopups, or both. maybe status changes to the player get displayed as both, whereas status changes on enemies are only indicated in their damagePopups.
also a cool idea would be to be able to break out of stun faster by mashing buttons. i'll see if that's possible, as it would require decrementing the duration of a Timer as it's running, which i'm not sure is possible. the reflexive solution is to get the time remaining on the stun timer, decrement it by an amount per input detected (like 0.1s), and then restart the timer at the new value, but i'm not sure if that's wise or efficient.