player can reduce stun time by mashing action buttons
action buttons are: attack, roll, examine, alchemy (VBFG). may include directional inputs as well.
i'm still not sure how to handle getting stunned/hit while casting. on one hand, i feel getting interrupted while casting would be too annoying for the kind of game i'm going for. on the other, i don't want the casting animation to be a temporary stun-immune period. so what i've done is added a check in action_finished, which is called when a spell ends. this checks for a "Stun" node on the Player, and if it's found, sets the player's state to Stun.
the reason i don't want formulas to be interrupted is because i want the decision to cast a spell to be a hard one - not hard as in difficult, but hard as in rigid. in other words, if you decide to cast a spell, that spell should come out - in other words, you are "committing" to the spell when you confirm it. you give up being able to move while casting, but can still take damage. that's the drawback.