many bugfixes (hopefully)
hopefully fixed an old bug where sometimes the sword hitbox would not be disabled properly if it was enabled and disabled on the same frame, allowing the player to walk around with an always-active sword hitbox.
stun icon added and updates with the correct time remaining after inputs while stunned. stun time is still hard-coded for the moment.