reworking frenzy status
before, frenzy was a Node permanently attached to the PlayerStats script, which had a method, activate(). when activate() was called, the Player's attack speed was multiplied by a factor of potency (2) over a period of duration (3). this also started an animation from the Player's FrenzyAnimationPlayer which modulated the Player's sprite to Red, and faded it out over the duration of the effect.
in addition to doubling attack speed, Frenzy also set the sword's knockback_vector to 1/10th of a normal attack, preventing enemies from being launched out of attack range. it also lowered the stamina costs of each attack. the way we used to handle this was with hard coded checks in the Player's attack functions:

		if PlayerStats.frenzy.active:
			swordHitbox.knockback_vector = input_vector / 10
		else:
			swordHitbox.knockback_vector = input_vector

and

	if PlayerStats.frenzy.active:
		stats.stamina -= 3
	else:
		stats.stamina -= 15

with the new status system, Statuses are added and removed from the Scenetree as children of the body that they are affecting, rather than permanently affixed nodes with activate functions.
so taking that into account, we must consider whether we want attacking during frenzy to go through a condition check (check if Frenzy is there), or perhaps have the existence of Frenzy alone modify the relevant values. what i mean is, for example, stamina drain/attack is hard-coded. it costs 15 stamina for a regular attack, and 3 stamina for an attack during frenzy. what if, instead of checking for frenzy and subtracting the appropriate amount, the existence of Frenzy changes a "stamina_cost" variable. we'll see what works.