further integration of status system
frenzy status implemented with the new system, but it still doesn't have a spell to apply it - 'T' is the test input to set the status.
it also doesn't have any behaviour for enemies, yet, since the values it modifies have to do with the player only, which are stamina cost and sword knockback vector. since enemy attacks are all governed by timers, it might be possible to have all those timers' durations set via variables, and then modify that variable when frenzied. say, if there is a "frenzy modulator" variable that defaults to 1 - all enemy timer durations are multiplied by this variable. if it gets set to 0.5, then in theory all enemy timers should run at have their default duration, cutting times for starting attacks & cooldowns by half, which would likely make the enemy "feel" a lot more aggressive. just a theory because it's not in practice yet. of course, "frenzy" could also simply be a Player-only status effect - i have no problem with that.
additionally, instead of adding statuses directly as children of the bodies they are applied to, i've added an extra Node2D per body called StatusDisplay. this is because different bodies/enemies will have different heights/sizes, and the status indicator should always display ABOVE their sprite. by having the StatusDisplay Node2D attached to every enemy, the position of the status effect indicator can be customized. in the case of the Crow, it even moves up and down depending on its animations.