working on Fury formula
the next challenge is to figure out how to affect the potencies/durations of status effects applied by formulas. formulas can level, and as they level, they grow stronger:
potency = (base_potency*formula_level_mod) * (magic_mod*magic_mod)
since there are 2 modifiers at play here, formula_level_mod (the modifier applied to the formula's base_potency depending on the formula's level) and magic_mod (the modifier, squared, which is applied to the result of the previous calculation), i think it may be worth thinking about having these 2 modifiers modulate different aspects of a spell. for healing & damage, both modifiers end up affecting the "potency", which ultimately determines how much health is healed or how much damage is dealt. but what about for formulas with durations, like status effects? it might be interesting to have the spell_level_mod extend the DURATION of a status effect, and have the magic_mod affect the POTENCY. in other words, the higher the spell level, the longer the effect lasts; the higher the magic stat, the more potent the effect. i will have to come up with new ways to calculate these values, but it's probably worth trying.
i was also thinking about maybe allowing the player to upgrade spells with "perks", perhaps at certain level breakpoints. a "perk" could be a choice between 2 augments to the spell, for example, would you rather extend the length of a spell's period, or augment the potency of a spell? or; would you like to expand the AoE radius of a spell, or increase its chance of inflicting a status effect? etc, etc. if this were the case, then EVERY value of a formula needs to be modifiable and needs to work properly no matter what values are assigned to it.