Fury formula added
it's nice to see the spellcasting system working, synthesizing all the different components of a spell is complicated but the system i have works... as long as all the values are properly set up beforehand.
next is to tackle the issue mentioned in the previous commit, which is to account for spell_level_mod and magic_mod in applying status effects.
the way statuses are applied is a bit more complicated than just a spell the deals damage. there are a number of additional steps involved.
right now, as previously mentioned, status values are hard-coded into their respective Scenes. ie. Poison.duration = 16 and Poison.potency = 1.
statuses are applied by calling the StatusHandler.apply_status(status, body) method, where status is a size two array that contains the name of the status as a string and the chance of infliction as a float. eg. status = ["Poison", 1.0] is an attempt to apply the Poison status with a 100% chance of infliction (before resists) on the body to which it's being applied.
the spellcasting system works as such: the FormulaBook is populated with formulas we currently have in our possession. when the "formula" key is hit, we look at what our current_selected_formula is and reference it - "it" being a resource file containing various parameters for the formula in question, including the ingredients required, as well as a NodePath to the formula scene itself. this is an important distinction: when we "cast", we are referencing the formula RESOURCE, inside of which is contained a path to the formula scene we should load.
now it gets complicated. it is these FORMULA SCENES whose values change depending on our PlayerStats. the calculations for potency are done inside the HITBOX of the formula. we want to change the values in our STATUS SCENES when our formula instantiates a status.

here is a clearer illustration of what currently happens:
FORMULA RESOURCE loads FORMULA SCENE which sets values via PLAYERSTATS

here is what we want to happen with a status formula:
FORMULA RESOURCE loads FORMULA SCENE which sets values via PLAYERSTATS which applies the appropriate status parameters via additional arguments to the STATUSHANDLER which then CREATES & APPLIES THE STATUS.

all that to say, i think the way to get this to work is to just code in additional arguments for the StatusHandler, which are to be the calculations performed in the status formula scene.