Fury level/magic stat modifies Frenzy status
because the status passed to the StatusHandler is an array, this means the array can have any number of indexes; it does not need to be fixed. therefore, statuses may have additional indexes if required. in the case of the Frenzy status, when the StatusHandler receives a Frenzy call, it applies status[2] to the status's duration and status[3] to the status's potency variables. the default values for duration and potency are 3.0 and 1.5, meaning a duration of 3 seconds and an attack speed multiplier of 1.5.
the duration of Frenzy is based off of the Fury formula level:
duration = 3.0 + formula_level_mod/2.0
so the ultimate duration of the status effect is 3 seconds + formula_level/2, aka 3 + 0.5s per formula_level (eg. level 1 = 3.5s; level 2 = 4s; level 10 = 8s)
the potency, aka the attack speed multiplier, is based off the magic mod:
potency = min(magic_mod, 2)
this, in effect, provides Frenzy with a 3.125% increase in attack speed PER point in Spirituality (magic stat), capping at a 100% increase. frenzy can never more than double your attack speed, which is why it picks the lesser between magic_mod & 2. potency will never go above 2.
at 4 magic, the multiplier is 1.125, which isn't much but is still barely noticeable. the attack speed bonus of frenzy should cap out (potency=2) at 32 Magic. of course these values and formulas will need adjusting.