working on creating new tilesets
so creating tilesets capable of auto-tiling is extremely complicated. many months ago, i created the tileset used in the Home map, and it took days of trial and error, just for that simple-looking wall+floor to auto-tile properly. the Godot tilemap system is notoriously bad. so instead of learning through experimentation and trial and error, i am starting from scratch, the basics, and studying/synthesizing everything i take in on the way, taking my time doing so instead of skipping steps in an attempt to get a result faster.
unfortunately, the Godot documentation is less than stellar. however, they do provide some template tilemaps and a guide on how to apply proper bitmasking to them. the TileMap node has a ton of parameters, and almost all of them are left unexplained. so this might be an arduous process, but i'm committed.
i have created some mountain "walls" to be used as borders, or edges to backgrounds, or part of the background itself. now, all i have to do is create a tileset that is capable of drawing this mountain border properly.