need more practice creating tilesets
the goal is not to become frustrated with time spent. i am making a lot of mistakes learning how to make a tileset from scratch, and it is a very time-ineffecient process. the part of the mountain tileset i created was an absolute pain due to how each "tile" is larger vertically than the cells they occupy, so putting together the tilemap was extremely unintuitive. i'm hoping to keep practicing with tilesets until i get used to making and using them. designing everything by hand is certainly possible, but designs can be created much faster with tilesets. you just spend a lot of time setting up and making them, but it makes the actual map design process much smoother.
again, the goal is not to become frustrated. so i will be making more tilesets to practice. the one i've created is a "wall" type, i guess, so i will try making a "floor" one next. hopefully i get the hang of it and develop an intuition for designing these things because as of now it is simply not much but a pain.
however, i think i will have to come across a fairly large problem, sooner or later. for the entirety of development, i have the Player and all world Objects as children of a YSort node, which itself is a child of the World. above the YSort node is a Node2D for the Map, which so far has just been things like floors and backgrounds. i'm not certain, but i think for Tilemaps to incorporate themselves into YSorting is to have objects be children directly of the Tilemap node itself. this is going to cause a very large number of headaches, since the Player is hard-referenced by getting its nodepath through the YSort node on an extremely large number of scripts.  the solution, i think, would be to create a generic, global function that gets a reference to the player, then change every other hard reference to the one that uses the global function - that way, the player's location in the SceneTree doesn't necessarily have to be "root/World/YSort/Player". the issue is going to be finding every single reference to the Player. it will likely take an entire day of just debugging. i have gotten better as a programmer since then, though, so maybe i can come up with a more eloquent solution.