collision/YSorting achieved on mountain Tileset
tilesets continue to be extremely time-consuming to work with, but i feel that if i can get a few "template" tilesets made of different types such as floor (64, 64), top-down wall in 3/4 perspective (64, 88), and tall walls (64, 184), they can be re-used quickly for any new art that is created. i do not want to go through this process for each tileset created.