beginning work on element system
after wracking my brain to come up with an elegant system for applying the appropriate factor/ratio to damage values, i have decided to hard-code damage ratios into arrays corresponding with the element used, where each index of the array contains the ratio to be applied against the target element.
right now, for testing, i have declared 10 elements: fire, ice, lightning, water, wind, earth, light, dark, verve, and entropy.
the calculate_element_ratio(source_element, target_element) function takes two arguments for the element of the attack and the element of the target.
we match the input source_element with the appropriate damage ratios. each element has a corresponding array of all the possible damage ratios for that element. for example:
FireMod = [0.5, 2, 1, 1, 1, 1, 1, 1, 1, 1]
the first index is the ratio to be used against Fire, the second index is the ratio to be used against Ice, etc.
i have created a standalone Element scene & script that holds the ratio values and the function to apply the appropriate one. i figure it could be integrated into the Global script.