tuning Elements and many fixes
element system seems to work OK. will need balancing as we move on but basically: weakness to an element multiplies the damage by 1.5, affinity with an element halves it. drained elemental damage shows up properly as healing.
i've also been adjusting the status system as there were a few fixes. one, when refreshing a buff, it would simply restart the buff with the values it was instantiated with. this means that if you leveled up your magic stat, or the formula that bestowed the buff, and then refreshed it, it would not use the new duration/potency. now it does.
additionally, the regen and frenzy buffs get their durations/potencies through some calculations. for frenzy:

duration = 3 + formula_level/2
potency = min(magic_mod, 2)

and for regen:

duration = 16 + (formula_level*2)
potency = int(formula_level+magic_mod*magic_mod)

because of this, buffs no longer have a default duration/potency. the values must always be calculated via stats. will have to come up with different calculation for enemy buffs; or, simply hard-code those values in.

i've also started to consider multiplayer, so hard references to the Player node need to be re-thought. for example, when a formula is cast, its 'player' variable is assigned to the body that cast the formula.

something that will need to be considered is the different player pouches and formulabooks, as right now the game/system is only designed to keep track of ONE pouch and ONE formulabook. it will need to be redesigned to create multiple pouch/formulabook resources and assign them to the player(s).