refactoring game systems to allow for multiplayer
right now, i'm just working on getting the game to run in its current state with a second player spawned. i have not yet updated the player 2 GUI, and have not yet incorporated discrete pouch/formulabook resources for the different players, so only player 1 can cast, for now.
so far i've rewritten a lot of the pause screen, level up screen, and various status effects to work with multiplayer.