zoning in multiplayer works
this is tricky. despite removing the old "items" from the game, the "inventory" resource remains. i decided to keep it in case i wanted to use it for something, like key items. because of this, however, it's caused some trouble in the player initializing/reinitialization methods. the "inventory" will only ever be a resource associated with Player 1.
when the game is run, a Player is spawned. when this happens, the player checks for Global.get_attribute("inventory"), which is the name we use to store data related to ingredients & formulas on-hand when changing zones - the attribute only gets assigned when we zone, not when we spawn. this is the key difference between "initialization" and "re-initialization".
the Player only ever gets "initialized" once when we run the game. he should never get initialized again. Player 2, however, can be initialized whenever he joins for the first time in a scene. it's probably not a problem for now since there is no real save system yet, but whenever a player gets initialized - as opposed to REinitialized - their inventory data is loaded from the disk. the only time this data gets saved is when it is created for the first time.
as a workaround, i have added the array player2_data to the GameManager. player2_data is an array with 2 indexes: 0 for pouch, and 1 for formulabook. this array gets set on initialization, reinitialization, and toggling multiplayer off (BackSlash). when a player zones, we check if GameManager.player2 != null, and if so, we pass player2's data as an array inside the Exit's new_inventory_2 variable. if player2 never gets initialized, GameManager.player2 will remain null; player2_data won't be populated, and the Global.goto_scene method will only use direction, location, and inventory as attributes - as opposed to direction, location, inventory and inventory_2.